+++ zoet-macro/src/zoet.rs
@@ -7,8 +7,8 @@ use crate::{
 use proc_macro2::{Ident, Span, TokenStream};
 use quote::{quote, ToTokens};
 use syn::{
-    parse::Parser, parse2, punctuated::Punctuated, spanned::Spanned, token, Attribute,
-    Generics, ImplItem, ImplItemMethod, Item, ItemFn, ItemImpl, Meta, NestedMeta, Signature, Type,
+    parse::Parser, parse2, punctuated::Punctuated, spanned::Spanned, token, Attribute, Generics,
+    ImplItem, ImplItemMethod, Item, ItemFn, ItemImpl, Meta, NestedMeta, Signature, Type,
 };
 
 fn trait_impl_from_fn(
@@ -20,6 +20,22 @@ fn trait_impl_from_fn(
     self_type: Option<&Type>,
     generics: &Generics,
 ) {
+    match *signature {
+        Signature { asyncness: Some(ref tok), .. } => abort!(
+            attr, "cannot apply to an async function";
+            help = tok.span() => "async function defined here"
+        ),
+        Signature { unsafety: Some(ref tok), .. } => abort!(
+            attr, "cannot apply to an unsafe function";
+            help = tok.span() => "unsafe function defined here"
+        ),
+        Signature { variadic: Some(ref tok), .. } => abort!(
+            tok, "cannot apply to a variadic function";
+            help = attr.span() => "variadic function defined here"
+        ),
+        _ => {}
+    };
+
     let ident = &signature.ident;
     let to_call = &self_type.map_or_else(
         || quote! { #ident },
@@ -59,6 +75,8 @@ fn trait_impl_from_fn(
 
                 let extra_attrs = quote! {
                     #[allow(
+                        // We don't/can't generate all optional methods for e.g. Iterator
+                        clippy::missing_trait_methods,
                         // Disable unhelpful "unnecessary structure name repetition" lint.
                         clippy::use_self,
                         // Disable unhelpful "unnecessary qualification" lint; paths *should* be
@@ -167,8 +185,7 @@ fn filter_attrs(attrs: &mut Vec<Attribute>) -> (Vec<Attribute>, Vec<Attribute>) 
                 //// It's probably not needless since we copy lifetimes into the generated trait impl
                 clippy::needless_lifetimes,
                 //// It's natural to give the wrapped inherent fn the same name as the trait impl fn
-                //// Disabled for now as this clippy lint seems a bit broken.
-                // clippy::same_name_method,
+                clippy::same_name_method,
             )]
         })
         .unwrap_or_abort();
@@ -217,8 +234,8 @@ fn zoet_inherent_impl(attr: &TokenStream, mut item_impl: ItemImpl) -> TokenStrea
     tokens
 }
 
-fn parse_nested_meta(nested_meta: NestedMeta) -> Result<(Ident, GenFn)> {
-    match nested_meta {
+fn parse_nested_meta(nested_meta: &NestedMeta) -> Result<(Ident, GenFn)> {
+    match *nested_meta {
         NestedMeta::Lit(ref lit) => Err(diagnostic_error!(lit, "literals are not valid here")),
         NestedMeta::Meta(ref meta) => match *meta {
             Meta::List(ref value) => Err(diagnostic_error!(value, "this does not take parameters")),
@@ -240,7 +257,7 @@ fn trait_fns(
 ) -> impl Iterator<Item = (Span, Result<(Ident, GenFn)>)> {
     nested_metas
         .into_iter()
-        .map(|nested_meta| (nested_meta.span(), parse_nested_meta(nested_meta)))
+        .map(|nested_meta| (nested_meta.span(), parse_nested_meta(&nested_meta)))
 }
 
 pub(crate) fn zoet(attr: &TokenStream, item: &TokenStream) -> TokenStream {
@@ -260,13 +277,7 @@ pub(crate) fn zoet(attr: &TokenStream, item: &TokenStream) -> TokenStream {
         parse2(item.clone()).unwrap_or_else(|error| abort!(item, "this is not an item: {}", error));
 
     match item {
-        Item::Fn(item_fn) => match item_fn {
-            ItemFn {
-                sig: Signature { variadic: None, .. },
-                ..
-            } => zoet_free_fn(attr, item_fn),
-            _ => abort!(item_fn, "cannot apply to variadic functions"),
-        },
+        Item::Fn(item_fn) => zoet_free_fn(attr, item_fn),
         Item::Impl(item_impl) => match item_impl {
             ItemImpl { trait_: None, .. } => zoet_inherent_impl(attr, item_impl),
             _ => abort!(item_impl, "can only apply to inherent impls"),