+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 2023-05-25: Release 0.1.11
+
+* There is extra checking on function signatures to give a useful diagnostic for `async` and `unsafe` fns.
+
+* The generated traits now also add `#[allow(clippy::missing_trait_methods)]` since they can only
+  ever implement the primary trait function and there's nothing the user of this macro can do to fix
+  this short of manually implementing the trait.
+
+* The wrapped function now also adds `#[allow(clippy::same_name_method)]` to suppress that
+  diagnostic. It is reasonable to name the function after the trait method, e.g. `#[zoet(Iterator)]
+  fn next(...)`
+
 ## 2022-10-29: Release 0.1.10
 
 * `#[doc_cfg]` attributes are no longer copied from the function to the derived impls as this was a